requirements.txt
Django==1.7.1
Pillow==3.2.0
Unidecode==0.04.16
Whoosh==2.6.0
beautifulsoup4==4.4.1
bleach==1.4
django-activelink==0.4
django-appconf==0.6
django-autocomplete-light==2.0.0a17
django-bootstrap3==4.8.2
django-celery==3.1.17
django-ckeditor==4.4.7
django-compressor==1.6
django-debug-toolbar==1.2.1
django-extensions==1.6.1
django-haystack==2.3.1
django-import-export==0.4.2
django-libsass==0.2
django-model-utils==2.2
django-modeltranslation==0.11
django-mptt==0.6.1
django-notifications-hq==0.6.2
django-redis-cache==1.5.3
django-social-auth==0.7.28
django-tables2==0.15.0
django-tagging==0.3.4
django-tastypie==0.13.3
elasticsearch==1.4.0
html5lib==0.999
ipython==4.1.1
httplib2==0.9
libsass==0.5.1
loremipsum==1.0.5
names==0.3.0
oauth2==1.5.211
opbeat==3.3
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2014.9
requests==2.9.1
shortuuid==0.4.2
six==1.7.3
sorl-thumbnail==11.12.1b
sqlparse==0.1.12
wsgiref==0.1.2
